refactor: Add a buffered StorageAPI helper
used for both chain storage and crank buffers
